# src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Polimoney Hub API",
    "version": "1.0.0",
    "description": "政治資金の透明化を推進する Polimoney プラットフォームの Hub API\n\n## 認証\n\n### API キー認証（公開 API）\n`X-API-Key` ヘッダーに API キーを設定してください。\n\n### Bearer トークン認証（管理者 API）\n`Authorization: Bearer <token>` ヘッダーに Supabase Auth のアクセストークンを設定してください。",
    "contact": {
      "name": "Polimoney Team",
      "url": "https://github.com/moai-redcap/polimoney_hub"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3722",
      "description": "ローカル開発環境"
    },
    {
      "url": "https://polimoney-hub.dd2030-topics.deno.net",
      "description": "本番環境"
    }
  ],
...
  "paths": {
...
    "/api/v1/polimoney/elections": {
      "get": {
        "tags": ["Polimoney"],
        "summary": "公開選挙一覧",
        "description": "収支データが公開されている選挙の一覧を取得します",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "api_version": { "type": "string", "example": "v1" },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ElectionSummary"
                      }
                    },
                    "total_count": { "type": "integer", "example": 5 }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/unlock-requests": {
      "post": {
        "tags": ["Unlock"],
        "summary": "ロック解除リクエスト作成",
        "description": "年度締め済み台帳のロック解除をリクエストします",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUnlockRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "作成成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/UnlockRequest" }
                  }
                }
              }
            }
          },
          "409": { "description": "既存の pending リクエストあり" }
        }
      },
      "get": {
        "tags": ["Unlock"],
        "summary": "ロック解除リクエスト一覧",
        "description": "ロック解除リクエストの一覧を取得します",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "ステータスでフィルタ",
            "schema": {
              "type": "string",
              "enum": ["pending", "approved", "rejected"]
            }
          },
          {
            "name": "ledger_id",
            "in": "query",
            "description": "台帳 ID でフィルタ",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/UnlockRequest" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/election-requests": {
      "get": {
        "tags": ["Admin"],
        "summary": "選挙リクエスト一覧",
        "description": "選挙登録リクエストの一覧を取得します",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "approved", "rejected", "needs_info"]
            }
          }
        ],
        "responses": {
          "200": { "description": "成功" }
        }
      }
    },
...
    "/api/admin/organization-requests": {
      "get": {
        "tags": ["Admin"],
        "summary": "団体リクエスト一覧",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "成功" }
        }
      }
    },
...
    "/api/admin/unlock-requests": {
      "get": {
        "tags": ["Admin"],
        "summary": "ロック解除リクエスト一覧（管理者）",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "成功" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API キー（開発用: dev_xxx, 本番用: prod_xxx）"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase Auth アクセストークン"
      }
    },
    "schemas": {
...
      "ElectionSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "type": { "type": "string" },
          "election_date": { "type": "string", "format": "date" },
          "district_id": { "type": "string", "nullable": true },
          "district_name": { "type": "string", "nullable": true }
        }
      },
...
      "CreateUnlockRequest": {
        "type": "object",
        "properties": {
          "ledger_id": { "type": "string", "format": "uuid" },
          "ledger_type": {
            "type": "string",
            "enum": ["election", "organization"]
          },
          "fiscal_year": {
            "type": "integer",
            "description": "organization の場合は必須"
          },
          "requested_by_user_id": { "type": "string", "format": "uuid" },
          "requested_by_email": { "type": "string", "format": "email" },
          "reason": { "type": "string", "minLength": 10 }
        },
        "required": [
          "ledger_id",
          "ledger_type",
          "requested_by_user_id",
          "requested_by_email",
          "reason"
        ]
      },
      "UnlockRequest": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "ledger_id": { "type": "string", "format": "uuid" },
          "ledger_type": {
            "type": "string",
            "enum": ["election", "organization"]
          },
          "fiscal_year": { "type": "integer", "nullable": true },
          "requested_by_user_id": { "type": "string", "format": "uuid" },
          "requested_by_email": { "type": "string", "format": "email" },
          "reason": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["pending", "approved", "rejected"]
          },
          "approved_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "approved_by": { "type": "string", "nullable": true },
          "unlock_expires_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "rejection_reason": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}